% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ST.Colours-package.R
\docType{package}
\name{ST.Colours-package}
\alias{ST.Colours-package}
\title{ST.Colours: The Sutton Trust Palettes}
\description{
Colour palette for the Sutton Trust that was built using the 'palettes' package. Optimised for accessibility.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/TolaniRuthOgun/ST.Colours}
  \item Report bugs at \url{https://github.com/TolaniRuthOgun/ST.Colours/issues}
}

}
\author{
\strong{Maintainer}: Tolani-Ruth Ogundamisi \email{ruth.ogundamisi.19@ucl.ac.uk}

}
\keyword{internal}
