% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ST.Colours-package.R, R/palettes.R
\docType{data}
\name{ST.Colours}
\alias{ST.Colours}
\alias{ST.Colours_discrete}
\alias{ST.Colours_sequential}
\title{Internal palettes methods}
\format{
\subsection{\code{ST.Colours}}{

An object of class \code{palettes_palette} with 18
colour palettes. Use \code{names(ST.Colours)} to return all palette names.
}

\subsection{\code{ST.Colours_discrete}}{

An object of class \code{palettes_palette} with
10 discrete colour palettes.
Use \code{names(ST.Colours_discrete)} to return all palette names.
}

\subsection{\code{ST.Colours_sequential}}{

An object of class \code{palettes_palette} with
8 sequential colour palettes.
Use \code{names(ST.Colours_sequential)} to return all palette names.
}
}
\source{
\url{https://github.com/TolaniRuthOgun/ST.Colours}
}
\usage{
ST.Colours

ST.Colours_discrete

ST.Colours_sequential
}
\description{
Palettes inspired by sunsets in the Canadian Prairies.
}
\examples{
# Get all palettes by name.
names(ST.Colours)

# Plot all palettes.
plot(ST.Colours)
}
\seealso{
\code{\link[=pal_palette]{pal_palette()}}, \code{\link[=pal_colour]{pal_colour()}}
}
\author{
\href{https://github.com/TolaniRuthOgun}{Tolani Ruth Ogundamisi}
}
\keyword{datasets}
\keyword{internal}
